Allow more overriding of options.
The invoice creation allowed for overriding options, except for verifyPos because it looked at the pre-merged array value. After being able to choose overrides on invoice creation, the validation function also needs to be able to read from the same override to be able to validate IPN hashes or not.